Modifed the way custom plugin widgets are handles and create a base class for custom widgets to inherit from.
